% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychUtils.R
\name{apaFormat}
\alias{apaFormat}
\title{apaFormat}
\usage{
apaFormat(val, OneMax = TRUE, Dec = 3, p = FALSE)
}
\arguments{
\item{val}{float}

\item{OneMax}{boolean}

\item{Dec}{integer}

\item{p}{boolean}
}
\value{
a formated value string
}
\description{
This function reformats a value to fit the apa style.
This means: removing leading zeros and rounding to 2 decimal places as a
default. If a p-value is inserted, a formatted p-vaue is returned.
}
\examples{
someValue <- apaFormat(someValue)
}
