% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{baseT}
\alias{baseT}
\title{baseT}
\usage{
baseT(t_object)
}
\arguments{
\item{t_object}{List}

\item{effect_num}{integer}

\item{return_name}{boolean}
}
\value{
a formated value string
}
\description{
This function takes in a base R t-test
and returns a formatted LaTeX t-/F-string fit for a paper.
}
