% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{chiString}
\alias{chiString}
\title{chiString}
\usage{
chiString(aov_object, effect, return_name = TRUE)
}
\arguments{
\item{effect}{string}

\item{affex}{mixed object}
}
\value{
a formated value string
}
\description{
This function takes in a afex mixed model and returns a
formatted chi-string fit for a paper.
}
