% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{descstring}
\alias{descstring}
\title{descstring}
\usage{
descstring(desc_object, effect_num = NULL, rt = TRUE)
}
\arguments{
\item{effect_num}{integer}

\item{lazy_object}{any lazystats object}

\item{effect_name}{string}
}
\value{
a formated value string
}
\description{
This function handles the formatting of any kind of descriptive
strings.
}
