% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{epsString}
\alias{epsString}
\title{epsString}
\usage{
epsString(aov_object, effect)
}
\arguments{
\item{aov_object}{List}

\item{effect}{string}
}
\value{
a formated value string
}
\description{
This function takes in a afex ANOVA and returns a formatted
epsilon for use in the reportAOV function.
}
