% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{lazydemographics}
\alias{lazydemographics}
\title{lazydemographics}
\usage{
lazydemographics(sample_orig = NULL, sample_final = NULL, full_text = TRUE)
}
\arguments{
\item{lazy_object}{any lazystats object}

\item{effect_num}{integer}

\item{effect_name}{string}
}
\value{
a formated value string
}
\description{
This function takes in one of my vpInfo objects and returns a
fully formatted participants section---or at least the part where
demographic information are reported.
}
