% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{lazyformat}
\alias{lazyformat}
\title{lazyformat}
\usage{
lazyformat(lazy_object, effect_num = 1, effect_name = NULL, convert = FALSE)
}
\arguments{
\item{lazy_object}{any lazystats object}

\item{effect_num}{integer}

\item{effect_name}{string}
}
\value{
a formated value string
}
\description{
This functions serves as the wrapper for all other objects.
It takes in any kind of object in the realm of lazystats and returns and apa
(7) formatted string.
}
