% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{lazySave}
\alias{lazySave}
\title{lazySave}
\usage{
lazySave(my_data, my_file, out_dir = OUT_DIR)
}
\arguments{
\item{timeVal}{an unfomarted float}
}
\value{
character string
}
\description{
lazily save a data.frame (or tibble) to a desired directory to
a ".csv" file
}
\examples{
lazySave(my_tibble)
}
