% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{lazyTime}
\alias{lazyTime}
\title{lazyTime}
\usage{
lazyTime(timeVal)
}
\arguments{
\item{timeVal}{an unfomarted float}
}
\value{
character string
}
\description{
This is a function that converts a base 10 (minute) time string
to a human-readable base 6 (60) h, m & s string.
}
\examples{
lazyTime(26.70683)
}
