% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{reportAOV}
\alias{reportAOV}
\title{reportAOV}
\usage{
reportAOV(aov_object, effect, return_name = TRUE)
}
\arguments{
\item{aov_object}{List}

\item{effect}{string}

\item{return_name}{boolean}
}
\value{
a formated value string
}
\description{
This function takes in a afex ANOVA and returns a formatted
LaTeX string fit for a paper.
}
