% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{reportChi}
\alias{reportChi}
\title{reportChi}
\usage{
reportChi(chi_object, effect, return_name = TRUE)
}
\arguments{
\item{effect}{string}

\item{return_name}{boolean}

\item{affex}{mixed object}
}
\value{
a formated value string
}
\description{
This function takes in a afex mixed model and returns a
formatted LaTeX string fit for a paper.
}
