% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{reportT}
\alias{reportT}
\title{reportT}
\usage{
reportT(t_object, effect_num)
}
\arguments{
\item{t_object}{List}

\item{effect_num}{integer}

\item{return_name}{boolean}
}
\value{
a formated value string
}
\description{
This function takes in a t-test from an emmeans comparisons
and returns a formatted LaTeX t-/F-string fit for a paper. It determines
whether the inserted object is a difference, or a difference of differences
and returns the appopriate statistic.
}
