% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{effString}
\alias{effString}
\title{effString}
\usage{
effString(aov_object, effect)
}
\arguments{
\item{aov_object}{List}

\item{effect}{string}
}
\value{
a formated value string
}
\description{
This function takes in a afex ANOVA and returns a formatted
effect string (ges) for use in the aovString function.
}
