% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{fString}
\alias{fString}
\title{fString}
\usage{
fString(aov_object, effect)
}
\arguments{
\item{aov_object}{List}

\item{effect}{string}
}
\value{
a formated value string
}
\description{
This function takes in a afex ANOVA and returns a formatted
F string [F(DFn, DFd) = f-value] for use in the aovString function
}
