% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychUtils.R
\name{pString}
\alias{pString}
\title{pString}
\usage{
pString(aov_object, effect)
}
\arguments{
\item{aov_object}{List}

\item{effect}{string}
}
\value{
a formated value string
}
\description{
This function takes in a afex ANOVA and returns a formatted
p string (p = p-value) for use in the aovString function
}
