% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazystats.R
\name{vpInfo}
\alias{vpInfo}
\title{lazydemographics}
\usage{
vpInfo(dat)
}
\arguments{
\item{dat}{dataset/tibble}
}
\value{
tibble
}
\description{
This function takes in one of my vpInfo objects and returns a
fully formatted participants section---or at least the part where
demographic information are reported.
It takes in a dataset, the analyzed variables are id, age (numeric),
gender (women, men, and non-binary individuals) and handedness (left vs. right).
}
